+++ app/Fields/Uitype/Date.php
@@ -58,10 +58,12 @@ class Date extends DateTime implements Uitype
     public function getFieldOptions() : array
     {
         return [
-            'repeated' => [
+            [
+                'key' => 'repeated',
+                'label' => trans('uccello::uitype.option.text.repeated'),
                 'type' => 'boolean',
-                'default_value' => false,
-            ],
+                'default' => false
+            ]
         ];
     }